pyproject.toml
[project]
name = "job-queue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.2.0",
]
